Provide dummy dispute coordinator by default.
