+++ node/service/Cargo.toml
@@ -173,3 +173,4 @@ try-runtime = [
 ]
 malus = ["full-node"]
 memory-stats = ["polkadot-overseer/memory-stats"]
+disputes = ["polkadot-node-core-dispute-coordinator/disputes"]
